refactor: 병렬 처리 중 중복 로직 제거
